# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "description": "错误详情。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "统一错误响应体。注册进 OpenAPI，使前端能生成对应类型。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExemptionCondition": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "$ref": "#/components/schemas/ExemptionField"
          },
          "value": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "field",
          "value"
        ],
        "title": "ExemptionCondition",
        "type": "object"
      },
      "ExemptionField": {
        "enum": [
          "expense_type",
          "invoice_type",
          "currency"
        ],
        "title": "ExemptionField",
        "type": "string"
      },
      "ExemptionGroup": {
        "additionalProperties": false,
        "properties": {
          "all": {
            "items": {
              "$ref": "#/components/schemas/ExemptionCondition"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "All",
            "type": "array"
          },
          "exemption_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Exemption Id",
            "type": "string"
          }
        },
        "required": [
          "exemption_id",
          "all"
        ],
        "title": "ExemptionGroup",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceDuplicateRuleDefinition": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "exemptions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExemptionGroup"
            },
            "maxItems": 100,
            "title": "Exemptions",
            "type": "array"
          },
          "kind": {
            "const": "invoice_duplicate",
            "title": "Kind",
            "type": "string"
          },
          "require_direct": {
            "default": false,
            "title": "Require Direct",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "kind"
        ],
        "title": "InvoiceDuplicateRuleDefinition",
        "type": "object"
      },
...
      "InvoiceTitleRuleDefinition": {
        "additionalProperties": false,
        "properties": {
          "allowed_titles": {
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Allowed Titles",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "exemptions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExemptionGroup"
            },
            "maxItems": 100,
            "title": "Exemptions",
            "type": "array"
          },
          "kind": {
            "const": "invoice_title",
            "title": "Kind",
            "type": "string"
          },
          "require_direct": {
            "default": false,
            "title": "Require Direct",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "allowed_titles"
        ],
        "title": "InvoiceTitleRuleDefinition",
        "type": "object"
      },
      "InvoiceTypeAllowance": {
        "additionalProperties": false,
        "properties": {
          "allowed_invoice_types": {
            "items": {
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Allowed Invoice Types",
            "type": "array"
          },
          "expense_type": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Expense Type",
            "type": "string"
          }
        },
        "required": [
          "expense_type",
          "allowed_invoice_types"
        ],
        "title": "InvoiceTypeAllowance",
        "type": "object"
      },
...
      "InvoiceTypeRuleDefinition": {
        "additionalProperties": false,
        "properties": {
          "allowances": {
            "items": {
              "$ref": "#/components/schemas/InvoiceTypeAllowance"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Allowances",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "exemptions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExemptionGroup"
            },
            "maxItems": 100,
            "title": "Exemptions",
            "type": "array"
          },
          "kind": {
            "const": "invoice_type",
            "title": "Kind",
            "type": "string"
          },
          "require_direct": {
            "default": false,
            "title": "Require Direct",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "allowances"
        ],
        "title": "InvoiceTypeRuleDefinition",
        "type": "object"
      },
...
      "LimitRuleDefinition": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "exemptions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExemptionGroup"
            },
            "maxItems": 100,
            "title": "Exemptions",
            "type": "array"
          },
          "kind": {
            "const": "limit",
            "title": "Kind",
            "type": "string"
          },
          "require_direct": {
            "default": false,
            "title": "Require Direct",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "thresholds": {
            "items": {
              "$ref": "#/components/schemas/LimitThreshold"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Thresholds",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "thresholds"
        ],
        "title": "LimitRuleDefinition",
        "type": "object"
      },
      "LimitThreshold": {
        "additionalProperties": false,
        "properties": {
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "expense_type": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Expense Type",
            "type": "string"
          },
          "max_amount": {
            "title": "Max Amount",
            "type": "string"
          }
        },
        "required": [
          "expense_type",
          "currency",
          "max_amount"
        ],
        "title": "LimitThreshold",
        "type": "object"
      },
...
      "RuleDefinition": {
        "discriminator": {
          "mapping": {
            "invoice_duplicate": "#/components/schemas/InvoiceDuplicateRuleDefinition",
            "invoice_title": "#/components/schemas/InvoiceTitleRuleDefinition",
            "invoice_type": "#/components/schemas/InvoiceTypeRuleDefinition",
            "limit": "#/components/schemas/LimitRuleDefinition",
            "timeliness": "#/components/schemas/TimelinessRuleDefinition"
          },
          "propertyName": "kind"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/LimitRuleDefinition"
          },
          {
            "$ref": "#/components/schemas/InvoiceTypeRuleDefinition"
          },
          {
            "$ref": "#/components/schemas/TimelinessRuleDefinition"
          },
          {
            "$ref": "#/components/schemas/InvoiceTitleRuleDefinition"
          },
          {
            "$ref": "#/components/schemas/InvoiceDuplicateRuleDefinition"
          }
        ]
      },
...
      "RuleVersionResponse": {
        "properties": {
          "config_fingerprint": {
            "title": "Config Fingerprint",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "definition": {
            "$ref": "#/components/schemas/RuleDefinition"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "rule_id",
          "version",
          "effective_from",
          "config_fingerprint",
          "definition",
          "created_by",
          "created_at"
        ],
        "title": "RuleVersionResponse",
        "type": "object"
      },
      "SaveRuleRequest": {
        "additionalProperties": false,
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/RuleDefinition"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "rule_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Rule Id",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "effective_from",
          "definition"
        ],
        "title": "SaveRuleRequest",
        "type": "object"
      },
      "SaveRuleResponse": {
        "properties": {
          "config_fingerprint": {
            "title": "Config Fingerprint",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "definition": {
            "$ref": "#/components/schemas/RuleDefinition"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reused_existing": {
            "title": "Reused Existing",
            "type": "boolean"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "rule_id",
          "version",
          "effective_from",
          "config_fingerprint",
          "definition",
          "created_by",
          "created_at",
          "reused_existing"
        ],
        "title": "SaveRuleResponse",
        "type": "object"
      },
...
      "TimelinessPolicy": {
        "additionalProperties": false,
        "properties": {
          "expense_type": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Expense Type",
            "type": "string"
          },
          "max_calendar_days": {
            "maximum": 3660.0,
            "minimum": 0.0,
            "title": "Max Calendar Days",
            "type": "integer"
          }
        },
        "required": [
          "expense_type",
          "max_calendar_days"
        ],
        "title": "TimelinessPolicy",
        "type": "object"
      },
      "TimelinessRuleDefinition": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "exemptions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExemptionGroup"
            },
            "maxItems": 100,
            "title": "Exemptions",
            "type": "array"
          },
          "kind": {
            "const": "timeliness",
            "title": "Kind",
            "type": "string"
          },
          "policies": {
            "items": {
              "$ref": "#/components/schemas/TimelinessPolicy"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Policies",
            "type": "array"
          },
          "require_direct": {
            "default": false,
            "title": "Require Direct",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "policies"
        ],
        "title": "TimelinessRuleDefinition",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "面向内部财务团队的费用报销预审系统",
    "title": "ExpenseGuard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/rules": {
      "get": {
        "operationId": "rules_list",
        "parameters": [
          {
            "in": "query",
            "name": "rule_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule Id"
            }
          },
          {
            "in": "query",
            "name": "latest_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Latest Only",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "eg_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eg Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleVersionResponse"
                  },
                  "title": "Response Rules List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List",
        "tags": [
          "rules"
        ]
      },
      "put": {
        "operationId": "rules_save",
        "parameters": [
          {
            "in": "cookie",
            "name": "eg_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Eg Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveRuleResponse"
                }
              }
            },
            "description": "OK"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Save",
        "tags": [
          "rules"
        ]
      }
    },
...
  }
}